+++ src/test/java/kitchenpos/fake/FakeOrderDao.java
@@ -1,7 +1,7 @@
 package kitchenpos.fake;
 
 import kitchenpos.dao.OrderDao;
-import kitchenpos.domain.Order;
+import kitchenpos.domain.order.Order;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
